requirements.txt
langchain
langchain-text-splitters
gdown
requests
openai
umap-learn
scikit-learn
langchain_community
tiktoken
langchain-openai
langchainhub
langchain-pinecone
pinecone
langchain-anthropic
PyMuPDF
pandas
numpy
matplotlib
streamlit
python-dotenv
